Test vim-colortemplate using Travis-CI
Use a Docker image based on Ubuntu 19.04, which includes vim 8.1.320.

Run a vim command-line that loads `templates/default_clone.colortemplate`
and tries to run `:Colortemplate` on it. Check that the resulting
`colors/default_clone.vim` was generated and that it is not an empty file.

Such a test can be configured to run on both the master branch and on open
PRs, it should make sure that the default template is kept working as the
code evolves.